+++ src/main/java/com/hedvig/notificationService/web/EventController.kt
@@ -7,7 +7,7 @@ import com.hedvig.notificationService.service.event.ContractRenewalQueuedEvent
 import com.hedvig.notificationService.service.event.QuoteCreatedEvent
 import com.hedvig.notificationService.service.event.EventRequest
 import com.hedvig.notificationService.service.event.StartDateUpdatedEvent
-import com.hedvig.notificationService.customerio.web.ChargeFailedEventDto
+import com.hedvig.notificationService.web.dto.ChargeFailedEventDto
 import com.hedvig.notificationService.service.request.EventRequestHandler
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PathVariable
