+++ src/test/java/com/hedvig/notificationService/service/event/HandleEventRequestsTest.kt
@@ -2,7 +2,6 @@ package com.hedvig.notificationService.service.event
 
 import com.hedvig.notificationService.service.request.EventRequestHandler
 import com.hedvig.notificationService.service.request.HandledRequestRepository
-import com.hedvig.notificationService.service.event.StartDateUpdatedEvent
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
