+++ src/test/java/com/hedvig/notificationService/service/event/WebEventRequestTest.kt
@@ -1,7 +1,6 @@
 package com.hedvig.notificationService.service.event
 
 import com.hedvig.notificationService.service.request.EventRequestHandler
-import com.hedvig.notificationService.service.event.StartDateUpdatedEvent
 import com.ninjasquad.springmockk.MockkBean
 import io.mockk.verify
 import org.assertj.core.api.Assertions
